requirements.txt
aiogram==3.15.0
aiohttp==3.11.11
sqlalchemy==2.0.36
aiosqlite==0.20.0
alembic==1.14.1
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn==0.34.0
jinja2==3.1.5
yookassa==3.4.0
Pillow==11.1.0
apscheduler==3.11.0
pydantic==2.10.4
httpx==0.28.1
certifi==2024.12.14
uuid6==2024.7.10
python-multipart==0.0.20
aiofiles==24.1.0